+++ docs/source/changelog.rst
@@ -2,6 +2,17 @@
 Changelog
 =========
 
+Upcoming version (not yet released)
+-----------------------------------
+
+Added
+^^^^^
+
+- Added reward term visualization to the native viewer (toggle with ``P``).
+
+.. figure:: _static/changelog/native_reward.png
+   :width: 80%
+
 Version 1.1.0 (February 12, 2026)
 ---------------------------------